Revert "Rename to zaptest/tlogger"
This reverts commit be2b5eab7ba2e0a9f947b715b42fcca03ce0157b.
